Dismantled QuestionPlayerStateService
